pyproject.toml
[tool.poetry]
name = "secrets-scanner"
version = "1.0.0"
description = "Enhanced Python secrets scanner for detecting hardcoded credentials"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "secrets_scanner"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^6.0"
detect-secrets = {version = "^1.4.0", optional = true}

[tool.poetry.extras]
detect-secrets = ["detect-secrets"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^22.1.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.931"

[tool.poetry.scripts]
secrets-scanner = "secrets_scanner.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py37', 'py38', 'py39', 'py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = ["detect_secrets.*"]
ignore_missing_imports = true
